YouTube_Summariser/requirements.txt
streamlit
google-generativeai
python-dotenv
youtube_transcript_api